(wip) refactor downloading system [no ci]
